Add thread processing and progress bar to recording import
Added the option to spawn processes to handle the recording import and
take into account when the script could run, the new option is called
DAYLIGHT.

At the same time the new implementation of the progress bar has been
added.

Add pull-request template

Update engine version in preparation for merge.